+++ src/main/app/common/court-calculations/courtDetermination.ts
@@ -23,6 +23,4 @@ export class CourtDetermination {
       }
     }
   }
-
-
 }
